+++ src/commonMain/kotlin/fr/acinq/lightning/channel/states/Syncing.kt
@@ -276,7 +276,7 @@ data class Syncing(val state: PersistedChannelState, val channelReestablishSent:
                             val closingTx = state.getMutualClosePublished(watch.tx)
                             val nextState = Closing(
                                 state.commitments,
-                                waitingSinceBlock = currentBlockHeight.toLong(),
+                                waitingSinceBlock = currentBlockHeight(),
                                 mutualCloseProposed = state.closingTxProposed.flatten().map { it.unsignedTx },
                                 mutualClosePublished = listOf(closingTx)
                             )
